Impl: Captura de KML, Fix: ICs Avulsos, Add args - Implementa o módulo earth.py em app/core - o bot responsável por fazer o download do kml associado ao IC - Remove auto-formatação de ICs avulsos (em app/gui/interface.py: InterfaceApp._acao_confirmar()) - o que resolve o problema de alguns ICs não estavam sendo triados. Agora o AutoTri só tria ICs com exatamente 15 caracteres. - Adiciona argumentos de controle de resiliência (--retry e --retry-delay) que alteram o número e delay entre as retrys do decorador @retry aplicado à fluxo_siatu (em app/pipeline/sistemas.py)

+++ app/utils/decorators.py
@@ -3,7 +3,7 @@ from functools import wraps
 from .logger import logger
 
 
-def retry(max_retries=3, delay=5, exceptions=(Exception,)):
+def retry(max_retries=3, delay=5.0, exceptions=(Exception,)):
     """
     Decorador para repetir a execução de uma função em caso de erro.
     Nesse contexto, não é necessário refresh do Selenium, porque
@@ -11,7 +11,7 @@ def retry(max_retries=3, delay=5, exceptions=(Exception,)):
 
     Args:
         max_retries (int): número máximo de tentativas.
-        delay (int): tempo (segundos) para esperar entre tentativas.
+        delay (float): tempo (segundos) para esperar entre tentativas.
         exceptions (tuple): exceções que devem disparar retry.
     """
